requirements.txt
jupyter
openai==0.28
flask
python-dotenv
serpapi
langchai
reportlab
unstructured
unstructured[pdf]